Provide SimRel dates in machine-readable form and add details generator
Provide SimRel dates as JSON because it's a commonly used and well
supported format, especially in the web area.

Add a generator for the Wiki details page for a release, that requires
the dates JSON file for the targeted release (i.e.
'<SimRel-name>_dates.json') to exist and can then be invoked from CLI:

java scripts/generateReleaseDetails.java <SimRel-name>

where <SimRel-name> is for example '2025-12', requiring to use Java-25
or later.